chel..

+++ lab4/Operation.java
@@ -1,8 +1,8 @@
 public class Operation {
-	public static byte[] randomArray(int size, int min, int max) {
-		byte[] array = new byte[size];
+	public static int[] randomArray(int size, int min, int max) {
+		int[] array = new int[size];
 		for (int i = 0; i < size; i++) {
-			array[i] = (byte)(Math.random()*(max-min+1)+min);
+			array[i] = (int)(Math.random()*(max-min+1)+min);
 		}
 		return array;
 	}
