+++ lab4/Sorter.java
@@ -1,9 +1,9 @@
 public interface Sorter {
 
-	void sort(byte[] array);
+	void sort(int[] array);
 
-	default void swapElements(byte[] array, int i, int j) {
-		byte temp = array[i];
+	default void swapElements(int[] array, int i, int j) {
+		int temp = array[i];
 		array[i] = array[j];
 		array[j] = temp;
 	}
